+++ mne_bids/dig.py
@@ -20,7 +20,7 @@ from mne_bids.config import (BIDS_IEEG_COORDINATE_FRAMES,
 from mne_bids.tsv_handler import _from_tsv
 from mne_bids.utils import (_extract_landmarks, _scale_coord_to_meters,
                             _write_json, _write_tsv)
-from mne_bids.path import get_entities_from_fname
+from mne_bids.path import get_entities_from_fname, BIDSPath
 
 
 def _handle_electrodes_reading(electrodes_fname, coord_frame,
@@ -291,8 +291,7 @@ def _coordsystem_json(*, raw, unit, hpi_coord_system, sensor_coord_system,
     _write_json(fname, fid_json, overwrite, verbose)
 
 
-def _write_dig_bids(electrodes_path, coordsystem_path, root,
-                    raw, datatype, overwrite=False, verbose=True):
+def _write_dig_bids(bids_path, raw, overwrite=False, verbose=True):
     """Write BIDS formatted DigMontage from Raw instance.
 
     Handles coordinatesystem.json and electrodes.tsv writing
@@ -300,17 +299,13 @@ def _write_dig_bids(electrodes_path, coordsystem_path, root,
 
     Parameters
     ----------
-    electrodes_path : BIDSPath
-        Filename to save the electrodes.tsv to.
-    coordsystem_path : BIDSPath
-        Filename to save the coordsystem.json to.
-    root : str | pathlib.Path
-        Path to the data directory
+    bids_path : BIDSPath
+        Path in the BIDS dataset to save the ``electrodes.tsv``
+        and ``coordsystem.json`` file for. ``datatype``
+        attribute must be ``eeg``, or ``ieeg``. For ``meg``
+        data, ``electrodes.tsv`` are not saved.
     raw : instance of Raw
         The data as MNE-Python Raw object.
-    datatype : str
-        Type of the data recording. Can be ``meg``, ``eeg``,
-        or ``ieeg``.
     overwrite : bool
         Whether to overwrite the existing file.
         Defaults to False.
@@ -332,6 +327,22 @@ def _write_dig_bids(electrodes_path, coordsystem_path, root,
     coord_frame_int = int(digpoint['coord_frame'])
     mne_coord_frame = MNE_FRAME_TO_STR.get(coord_frame_int, None)
     coord_frame = MNE_TO_BIDS_FRAMES.get(mne_coord_frame, None)
+
+    # create electrodes/coordsystem files using a subset of entities
+    # that are specified for these files in the specification
+    coord_file_entities = {
+        'root': bids_path.root,
+        'datatype': bids_path.datatype,
+        'subject': bids_path.subject,
+        'session': bids_path.session,
+        'acquisition': bids_path.acquisition,
+        'space': bids_path.space
+    }
+    datatype = bids_path.datatype
+    electrodes_path = BIDSPath(**coord_file_entities, suffix='electrodes',
+                               extension='.tsv')
+    coordsystem_path = BIDSPath(**coord_file_entities, suffix='coordsystem',
+                                extension='.json')
 
     if verbose:
         print("Writing electrodes file to... ", electrodes_path)